# Change `Zip` data structure to a flat one