Instead of creating a `zip` of more than two iterators by forming a tree structure, store all iterators in a tuple. It should be possible to get rid of `Zip1` and `Zip2`, but I wasn't sure of a good benchmark to quickly check locally whether the more general code was sub-optimal in the short case, hence WIP—ideas welcome!

Fixes #26765.