@martinholters 2018-09-18
@Keno you have implemented this elaborate dance:
https://github.com/JuliaLang/julia/blob/3a15e6bd5b13174de6958c875a41bff89cf117d1/base/iterators.jl#L295-L316

IIUC, you first iterate those iterators where `isdone` yielded `missing` because that has a change to bail out early, i.e. if the first iterator gives `nothing` there is no need to even call `iterate` on the second one. Do you have an example where that makes a tangible difference? (I have not replicated this in full for the `Zip` iterator here, yet,  and a test case would help me understand the issue when doing so.)