@haampie 2018-09-18
About the `isdone` stuff: if iterator 1 is not exhausted but stateful and iterator 2 is exhausted but stateful, then you should not advance iterator 1. An example is where you have one iterator reading lines from a file, zipped with a range `1:5` or something -- iterator 1 should not end up reading line number 6.