@martinholters 2018-09-19
Stateful iterators! Of course! <head against forehead>

Thanks for the pointers!