@martinholters 2018-09-21
i did some benchmarking `collect`ing `zip`ped `UnitRange`s as those should contribute relatively little runtime themselves. Here is what I got:

`collect(...)`|master|this PR|
---|---|---
`zip(1:1)`|28.990 ns|28.008 ns
`zip(1:1,1:1)`|28.710 ns|30.705 ns
`zip(1:1,1:1,1:1)`|32.954 ns|34.288 ns
`zip(1:1,1:1,1:1,1:1)`|304.869 ns|118.260 ns
`zip(1:1000)`|1.090 μs|1.097 μs
`zip(1:1000,1:1000)`|1.767 μs|1.781 μs
`zip(1:1000,1:1000,1:1000)`|2.305 μs|2.351 μs
`zip(1:1000,1:1000,1:1000,1:1000)`|757.447 μs|4.971 μs

The only regressions above measurement noise here are for `zip(1:1,1:1)` and `zip(1:1,1:1,1:1)`, so apparently the setup time in these cases is slightly higher.  I'm inclined not to worry about this. On the other hand, I'm surprised by the improvement for 4 `zip`ped iterators, where the old `Zip` is still type-stable (at least on the surface).

However, I have no idea whether these benchmarks are relevant, and BaseBenchmarks do not include any direct benchmarks of `zip` performance. So I'm all ears for proposals here.

That said, I think the code is now in a state where review makes sense (unless CI finds something).