@haampie 2018-09-21
I think the last row actually hits a type instability:

```julia
f() = collect(zip(1:1000,1:1000,1:1000,1:1000))
@btime f()
```

gives `1.254 ms (16439 allocations: 678.52 KiB)` on Julia 1.0 and `6.328 μs (2 allocations: 31.33 KiB)` on this branch.