@martinholters 2018-09-24
AV 64bit says:
```
Error in testset bigint:
Test Failed at C:\projects\julia\julia-2baf73b8f9\share\julia\test\bigint.jl:422
  Expression: cmp(big(x), x) === cmp(x, big(x)) === 0
   Evaluated: -1 === 1 === 0
ERROR: LoadError: Test run finished with errors
```
I think is due to a bug in `big`:
```
julia> big(Int8(-128))
-18446744073709551488
```
(`x` is chosen randomly in the failed test above, so may just have hit this case.)