@martinholters 2018-09-25
Hm, master has improved its handling of the old `Zip` implementation:

`collect(...)`|master|this PR|
---|---|---
`zip(1:1)`|27.867 ns|28.849 ns
`zip(1:1,1:1)`|28.878 ns|30.777 ns
`zip(1:1,1:1,1:1)`|32.988 ns|34.253 ns
`zip(1:1,1:1,1:1,1:1)`|115.974 ns|121.897 ns
`zip(1:1000)`|1.069 μs|1.069 μs|1.069 μs
`zip(1:1000,1:1000)`|1.781 μs|1.613 μs
`zip(1:1000,1:1000,1:1000)`|2.413 μs|2.367 μs
`zip(1:1000,1:1000,1:1000,1:1000)`|3.204 μs|4.749 μs

Inferability is still slightly better:
```julia
julia> x = zip(1:1, ["a"], (1.0,), Base.OneTo(1), Iterators.repeated("a"), 1.0:0.2:2.0,
               (1 for i in 1:1), Iterators.Stateful(["a"]), (1.0 for i in 1:2, j in 1:3), 1);

julia> z = Iterators.filter(x -> x[1] == 1, x);

julia> Base.return_types(first, Tuple{typeof(z)})
1-element Array{Any,1}:   # this PR
 Tuple{Int64,String,Float64,Int64,String,Float64,Int64,String,Float64,Int64}
1-element Array{Any,1}:   # master and 1.0
 Tuple{Int64,Any,Float64,Any,String,Any,Int64,String,Float64,Int64}
```

But I'm not sure the balance struck by the PR in its current form is better than master. If anyone wants to take a stab at tweaking this further, welcome! I want have time to work on this for at least one week.