@haampie 2018-09-25
Which version exactly? For reference:

```julia
julia> versioninfo()
Julia Version 1.0.0
Commit 5d4eaca0c9 (2018-08-08 20:58 UTC)
Platform Info:
  OS: Linux (x86_64-pc-linux-gnu)
  CPU: Intel(R) Xeon(R) CPU E3-1230 v5 @ 3.40GHz
  WORD_SIZE: 64
  LIBM: libopenlibm
  LLVM: libLLVM-6.0.0 (ORCJIT, skylake)
julia> f() = collect(zip(1:1000,1:1000,1:1000,1:1000));
julia> @benchmark f()
BenchmarkTools.Trial: 
  memory estimate:  678.52 KiB
  allocs estimate:  16439
  --------------
  minimum time:     619.442 μs (0.00% GC)
  median time:      625.912 μs (0.00% GC)
  mean time:        681.834 μs (5.98% GC)
  maximum time:     31.411 ms (97.53% GC)
  --------------
  samples:          7326
  evals/sample:     1
```