@martinholters 2018-09-25
Ah, it's a question of compilation (inference) order. I created the data for above tables with
```julia
for N in (1,1000), M in (1,2,3,4)
    @btime collect(zip($(Iterators.repeated(1:N, M)...,)...))
end
```
That gives the very fast timings on 1.0, also for timing `f()` in the same session later on. Reversing the order of values for `M` (number of `zip`ped iterators) and using a fresh session gives the bad timing again, also on master. This PR make the performance independent of the compilation/inference order (which is good), but the performance is a bit worse than the best that was possible before (which is not so good). Why can't I have my cake and eat it too?