@oscardssmith 2018-09-25
Why would performance be different for different orders? That sounds like a bug which should be fixed.