@martinholters 2018-09-25
Inference will limit its work on recursive functions (it has to to avoid working infinitely). However, if it hits a previously computed result before it reaches its limit, it will use that. That's why running inference on increasing problem sizes often produces better results than immediately running it on the largest problem size.