@haampie 2018-09-28
So, what is left to do in this PR? Is the only issue that we have to figure out why this PR is somewhat slower than master when inference would work well?