@martinholters 2018-10-04
A little optimization further, this is where I am now:

`collect(...)` | master (good) | master (bad) | this PR
---|---|---|---
`zip(1:1)` |   28.920 ns |   28.780 ns | 28.527 ns
`zip(1:1, 1:1)` |   29.621 ns |   29.581 ns | 29.621 ns
`zip(1:1, 1:1, 1:1)` |   32.706 ns |   32.814 ns | 33.338 ns
`zip(1:1, 1:1, 1:1, 1:1)` |   115.395 ns |   286.117 ns | 119.554 ns
`zip(1:1000)` |   1.067 μs |   1.062 μs | 1.069 μs
`zip(1:1000, 1:1000)` |   1.446 μs |   1.704 μs | 1.774 μs
`zip(1:1000, 1:1000, 1:1000)` |   2.351 μs |   2.328 μs | 2.297 μs
`zip(1:1000, 1:1000, 1:1000, 1:1000)` |   3.038 μs |   717.706 μs | 3.117 μs

The two "master" columns are for different evaluation order, either allowing or not allowing successful inference of the four-iterator case. That's probably as good as it gets.

It's noteworthy that there is a "jump" between three and four zipped iterators even for this PR (also showing additional allocations). That is apparently due to missing specialization here:
https://github.com/JuliaLang/julia/blob/60cff80396d39ace10175bd1505630513c7aa844/base/array.jl#L533
If I replace that one with
```julia
_similar_for(c::AbstractArray, ::Type{T}, itr, ::HasShape) where {T} = similar(c, T, axes(itr))
```
I get

`collect(...)` | master (good) | master (bad) | this PR
---|---|---|---
`zip(1:1, 1:1, 1:1, 1:1)` |  37.033 ns | 224.877  ns | 38.090 ns
`zip(1:1000, 1:1000, 1:1000, 1:1000)` |  2.476 μs |  694.169 μs | 2.468 μs

So that might be a worthwhile change, but orthogonal to the changes herein.