@haampie 2018-10-04
What confuses me is the result for zipping 2 iterators in "master (good)". Since that one should be inferable on master, why is it faster on "master (good")? Is it a typo?